+++ decidim-core/lib/decidim/core/engine.rb
@@ -200,7 +200,7 @@ module Decidim
 
       initializer "decidim.content_processors" do |_app|
         Decidim.configure do |config|
-          config.content_processors += [:user]
+          config.content_processors += [:user, :hashtag]
         end
       end