+++ decidim-core/lib/decidim/core/test/factories.rb
@@ -16,6 +16,10 @@ FactoryBot.define do
     "#{Faker::Lorem.characters(rand(1..10))}_#{n}"
   end
 
+  sequence(:hashtag_name) do |n|
+    "#{Faker::Lorem.characters(rand(1..10))}_#{n}"
+  end
+
   sequence(:email) do |n|
     "user#{n}@example.org"
   end
@@ -67,8 +71,8 @@ FactoryBot.define do
     official_img_header { Decidim::Dev.test_file("avatar.jpg", "image/jpeg") }
     official_img_footer { Decidim::Dev.test_file("avatar.jpg", "image/jpeg") }
     official_url { Faker::Internet.url }
-    highlighted_content_banner_enabled false
-    enable_omnipresent_banner false
+    highlighted_content_banner_enabled { false }
+    enable_omnipresent_banner { false }
     tos_version { Time.current }
 
     trait :with_tos do
@@ -81,13 +85,13 @@ FactoryBot.define do
 
   factory :user, class: "Decidim::User" do
     email { generate(:email) }
-    password "password1234"
+    password { "password1234" }
     password_confirmation { password }
     name { generate(:name) }
     nickname { generate(:nickname) }
     organization
     locale { organization.default_locale }
-    tos_agreement "1"
+    tos_agreement { "1" }
     avatar { Decidim::Dev.test_file("avatar.jpg", "image/jpeg") }
     personal_url { Faker::Internet.url }
     about { Faker::Lorem.paragraph(2) }
@@ -104,7 +108,7 @@ FactoryBot.define do
     end
 
     trait :deleted do
-      email ""
+      email { "" }
       deleted_at { Time.current }
     end
 
@@ -147,7 +151,7 @@ FactoryBot.define do
     organization
 
     transient do
-      users []
+      users { [] }
     end
 
     trait :verified do
@@ -174,7 +178,7 @@ FactoryBot.define do
   end
 
   factory :identity, class: "Decidim::Identity" do
-    provider "facebook"
+    provider { "facebook" }
     sequence(:uid)
     user
     organization { user.organization }
@@ -191,7 +195,7 @@ FactoryBot.define do
     end
 
     trait :pending do
-      granted_at nil
+      granted_at { nil }
     end
   end
 
@@ -245,7 +249,7 @@ FactoryBot.define do
 
     name { Decidim::Faker::Localized.sentence(3) }
     participatory_space { create(:participatory_process, organization: organization) }
-    manifest_name "dummy"
+    manifest_name { "dummy" }
     published_at { Time.current }
 
     trait :unpublished do
@@ -335,7 +339,7 @@ FactoryBot.define do
   factory :report, class: "Decidim::Report" do
     moderation
     user { build(:user, organization: moderation.reportable.organization) }
-    reason "spam"
+    reason { "spam" }
   end
 
   factory :impersonation_log, class: "Decidim::ImpersonationLog" do
@@ -441,5 +445,10 @@ FactoryBot.define do
     manifest_name { :my_manifest }
     weight { 1 }
     published_at { Time.current }
+  end
+
+  factory :hashtag, class: "Decidim::Hashtag" do
+    name { generate(:hashtag_name) }
+    organization
   end
 end