+++ decidim-proposals/app/commands/decidim/proposals/update_proposal.rb
@@ -48,9 +48,12 @@ module Decidim
       attr_reader :form, :proposal, :current_user, :attachment
 
       def update_proposal
+        parsed_title = Decidim::ContentProcessor.parse_with_processor(:hashtag, form.title, current_organization: form.current_organization).rewrite
+        parsed_body = Decidim::ContentProcessor.parse_with_processor(:hashtag, form.body, current_organization: form.current_organization).rewrite
+
         @proposal.update!(
-          title: form.title,
-          body: form.body,
+          title: parsed_title,
+          body: parsed_body,
           category: form.category,
           scope: form.scope,
           address: form.address,